Compiles under linux now
<==== Mercurial data ====>
Branch: master

+++ common/src/utils/color_space/image_correction.cpp
@@ -14,9 +14,9 @@ float ImageCorrectionResult::calcGamma(int leftPos, int rightPos, int pixels) co
         sum += hystogram[median];
 
     // 2. calc gamma
-    float curValue = (median - leftPos) / float(rightPos-leftPos+1);
-    float recValue = float(rightPos+leftPos) / 2.0 / 256.0;
-    return qBound(0.5f, log(recValue) / log(curValue), 2.0f);
+    qreal curValue = (median - leftPos) / qreal(rightPos-leftPos+1);
+    qreal recValue = qreal(rightPos+leftPos) / 2.0 / 256.0;
+    return qBound(0.5, (qreal)log(recValue) / log(curValue), 2.0);
 }
 
 void ImageCorrectionResult::processImage( quint8* yPlane, int width, int height, int stride,
